Fix crash on first startup in EEA
Solves an issue in which, with a system language whose country code is
a country within the European Economic Area (extension of the EU), the
browser may crash in template_url_prepopulate_data.cc with this error:
`Check failed: t_urls.size() <= kMaxEeaPrepopulatedEngines (10 vs. 8)`

Discovered by a user in the CalyxOS Development channel. Does not affect
Cromite itself, but would impact forks which do not include
"Enable-search-engine-settings-desktop-ui.patch".

Issue: calyxos#2636